% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovitzRRR.R
\name{MarkovitzRRRAlt}
\alias{MarkovitzRRRAlt}
\title{Compute Markovitz RRR}
\usage{
MarkovitzRRRAlt(
  returns,
  penalty_parameters,
  max_iter = 10000,
  tolerance = -1,
  check_arguments = TRUE
)
}
\arguments{
\item{returns}{\verb{n_observations x n_returns}-dimensional matrix of test asset
excess returns.}

\item{penalty_parameters}{in principle this is a \code{n_parameters}-dimensional
vector of penalty
parameter values from smallest to largest; for now it is just a positive number.}

\item{max_iter}{numeric solver parameter indicating the maximum number of
iterations. Default is \code{10000}.}

\item{tolerance}{numeric tolerance check for \verb{||X_k+1 - X_k||_F^2 / N^2}.
If \code{tolerance <= 0}, no check is performed. Default is \code{-1.}, for no .}

\item{check_arguments}{boolean \code{TRUE} if you want to check function arguments;
\code{FALSE} otherwise. Default is \code{TRUE}.}
}
\value{
a list containing
}
\description{
Computes Markovitz RRR
}
