% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovitzRRR.R, R/utils.R
\name{MarkovitzRRR}
\alias{MarkovitzRRR}
\alias{PlotMarkovitzRRRObjective}
\title{Compute Markovitz RRR}
\usage{
MarkovitzRRR(
  returns,
  penalty_parameter,
  initial_solution = matrix(0, 0, 0),
  penalty_type = "d",
  step_size_type = "d",
  step_size_constant = 0.001,
  max_iter = 10000,
  tolerance = -1,
  check_arguments = TRUE
)

PlotMarkovitzRRRObjective(results)
}
\arguments{
\item{returns}{\verb{n_observations x n_returns}-dimensional matrix of centred(!)
test asset excess returns.}

\item{penalty_parameter}{a positive number indicating the penalty parameter.}

\item{initial_solution}{\verb{n_returns x n_returns}-dimensional matrix of initial
hedging weights. Defaults is a hollow matrix with \code{1/N} as off-diagonal elements.}

\item{penalty_type}{character indicating the type of penalty function: \code{'d'}
for default, i.e., penalty given by \verb{||RX||_*}; \code{'a'} for alternative, i.e.,
penalty given by \verb{||X||_*}. Default is \code{'d'}.}

\item{step_size_type}{character indicating the type of step size:
\code{'d'} for default, i.e., not summable vanishing:
\code{step_size_constant / sqrt(iter + 1)};
\code{'s'} for square summable but not summable: \code{step_size_constant / (iter + 1)};
\code{'l'} for constant step length: \verb{step_size_constant / ||subgradient||_F}.
\code{'p'} for modified Polyak:
\verb{(step_size_constant + objective_iter - min\{objective_k | k=0,...,iter\}) / ||subgradient)||_F}.
\code{'c'} for constant step size: \code{step_size_constant}.
Default is \code{'d'}.}

\item{step_size_constant}{numeric constant determining the step size. Default
is \code{1.e-3}}

\item{max_iter}{numeric solver parameter indicating the maximum number of
iterations. Default is \code{10000}.}

\item{tolerance}{numeric tolerance check for \verb{||X_k+1 - X_k||_F^2 / N^2}.
If \code{tolerance <= 0}, no check is performed. Default is \code{-1.}, for no .}

\item{check_arguments}{boolean \code{TRUE} if you want to check function arguments;
\code{FALSE} otherwise. Default is \code{TRUE}.}

\item{results}{list containing a numeric vector \code{objective} with the
markovitzRRR objective function values at each iteration.}
}
\value{
a list containing
}
\description{
Computes Markovitz RRR

Plots the  MarkovitzRRR objective function value at each iteration
}
