% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovitzRRR.R
\name{MarkovitzRRR}
\alias{MarkovitzRRR}
\title{Compute Markovitz RRR}
\usage{
MarkovitzRRR(
  returns,
  penalty_parameters,
  max_iter = 10000,
  step_size_type = "s",
  step_size_constant = 0.001,
  tolerance = 1e-15,
  check_arguments = TRUE
)
}
\arguments{
\item{returns}{\verb{n_observations x n_returns}-dimensional matrix of test asset
excess returns.}

\item{penalty_parameters}{\code{n_parameters}-dimensional vector of penalty
parameter values from smallest to largest.}

\item{max_iter}{numeric solver parameter indicating the maximum number of
iterations. Default is \code{2000}}

\item{step_size_type}{character indicating the type of step size: \code{'s'} for
square summable but not summable and \code{'c'} for constant step size, any other
character gives a summable vanishing step size. Default is \code{'s'}.}

\item{step_size_constant}{numeric constant determining the step size. Default
is \code{1.e-3}}

\item{tolerance}{numeric tolerance check for \verb{||X_k+1 - X_k||_F^2 / N^2}.
Default is 1.e-15.}

\item{check_arguments}{boolean \code{TRUE} if you want to check function arguments;
\code{FALSE} otherwise. Default is \code{TRUE}.}
}
\value{
a list.
}
\description{
Computes Markovitz RRR
}
